questao2_backend/projeto-2---backend-tubarao-main/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-environ==0.10.0
django-filter==23.1
djangorestframework==3.14.0
idna==3.4
Markdown==3.4.3
pytz==2023.3
requests==2.29.0
sqlparse==0.4.4
urllib3==1.26.15
